Replaced business logic from some Controller methods to Service methods. Tested changed service methods. Tested Controller methods. Added new Exceptions for ExceptionHandling.

+++ src/main/java/clearsolutions/com/javatestassignment/controller/UserController.java
@@ -3,15 +3,11 @@ package clearsolutions.com.javatestassignment.controller;
 import clearsolutions.com.javatestassignment.model.User;
 import clearsolutions.com.javatestassignment.service.UserService;
 import jakarta.validation.Valid;
-import org.springframework.hateoas.EntityModel;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
 import org.springframework.web.bind.annotation.*;
-import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
-import java.lang.reflect.Field;
-import java.net.URI;
 import java.time.LocalDate;
 import java.util.List;
 
@@ -31,61 +27,53 @@ public class UserController {
     }
 
     @GetMapping("/user/{id}")
-    public EntityModel<User> retrieveUserById(@PathVariable Long id) {
+    public ResponseEntity<Object> retrieveUserById(@PathVariable Long id) {
         User user = userService.findById(id);
         if (user == null)
-            return null;
-        return EntityModel.of(user);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).build();
+        return ResponseEntity.status(HttpStatus.OK).body(user);
     }
 
     @PostMapping("/user")
     public ResponseEntity<?> createUser(@Valid @RequestBody User user) {
         user.setId(null);
         User savedUser = userService.save(user);
-        URI location = ServletUriComponentsBuilder.fromCurrentRequest()
-                .path("/{id}")
-                .buildAndExpand(savedUser.getId())
-                .toUri();
-        return ResponseEntity.created(location).build();
+        return ResponseEntity.status(HttpStatus.CREATED).body(savedUser);
     }
 
     @DeleteMapping("/user/{id}")
     public ResponseEntity<Void> deleteUser(@PathVariable Long id) {
-        userService.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        if(userService.findById(id) == null) {
+            return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+        }
+        else {
+            userService.deleteById(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        }
     }
 
     @PutMapping("/user/{id}")
     public ResponseEntity<User> updateUser(@PathVariable Long id, @Valid @RequestBody User user) {
-        userService.findById(id);
+        if(userService.findById(id) == null)
+            return ResponseEntity.notFound().build();
         user.setId(id);
-        userService.save(user);
+        if(userService.save(user) == null)
+            return ResponseEntity.badRequest().build();
         return ResponseEntity.ok(user);
     }
 
     @PatchMapping("/user/{id}")
-    public ResponseEntity<User> updateUserPartially(@PathVariable Long id, @RequestBody User user) throws IllegalAccessException, NoSuchFieldException {
-        User foundUser = userService.findById(id);
-        for (Field field : user.getClass().getDeclaredFields()) {
-            field.setAccessible(true);
-            String name = field.getName();
-            Object value = field.get(user);
-
-            if (null != value) {
-                Field destField = foundUser.getClass().getDeclaredField(name);
-                destField.setAccessible(true);
-                destField.set(foundUser, value);
-            }
-        }
-        userService.save(foundUser);
+    public ResponseEntity<User> updateUserPartially(@PathVariable Long id, @RequestBody User user) {
+        User foundUser = userService.updateUserPartially(id, user);
+        if(foundUser == null)
+            return ResponseEntity.notFound().build();
+        if(userService.save(foundUser) == null)
+            return ResponseEntity.badRequest().build();
         return ResponseEntity.ok(foundUser);
     }
 
     @GetMapping("/users/{from}/{to}")
-    public List<User> retrieveUsers(@PathVariable LocalDate from, @PathVariable LocalDate to) {
-        if(from.isAfter(to))
-            throw new IllegalArgumentException(from + " is after " + to);
-        else
-            return userService.findAllUsersInRange(from, to);
+    public List<User> retrieveUsersInRange(@PathVariable LocalDate from, @PathVariable LocalDate to) {
+        return userService.findAllUsersInRange(from, to);
     }
 }
